refactor: move project item to data
