+++ src/app/resume-page/experience-section/experience/experience-to-contents.spec.ts
@@ -2,7 +2,7 @@ import { TextContentComponent } from '../../chipped-content/text-content/text-co
 import { experienceToContents } from './experience-to-contents'
 import { makeExperience } from '../../data/__tests__/make-experience'
 import { ExperienceHighlightsComponent } from './experience-highlights/experience-highlights.component'
-import { ProjectItem } from '../../projects-section/project-item/project-item'
+import { ProjectItem } from '../../data/project-item'
 import { makeProjectItem } from '../../projects-section/__tests__/make-project-item'
 import { ExperienceTechComponent } from './experience-tech/experience-tech.component'
 import { ChippedContent } from '../../chipped-content/chipped-content'
