+++ src/app/resume-page/projects-section/adapt-json-resume-project.spec.ts
@@ -4,7 +4,7 @@ import {
   InvalidStackValueError,
 } from './adapt-json-resume-project'
 import { MockProvider } from 'ng-mocks'
-import { Stack } from './project-item/project-item'
+import { Stack } from '../data/project-item'
 import { serviceTestSetup } from '@/test/helpers/service-test-setup'
 import {
   RELATIVIZE_PRODUCTION_URL,
