+++ src/app/resume-page/projects-section/project-item/project-item-to-contents.ts
@@ -1,5 +1,5 @@
 import { ChippedContent } from '../../chipped-content/chipped-content'
-import { ProjectItem } from './project-item'
+import { ProjectItem } from '../../data/project-item'
 import { TextContentComponent } from '../../chipped-content/text-content/text-content.component'
 import { ProjectItemTechnologiesComponent } from './project-item-technologies/project-item-technologies.component'
 import { isNotUndefined } from '@/common/is-not-undefined'
