+++ src/app/resume-page/projects-section/project-item/project-item.component.spec.ts
@@ -9,7 +9,7 @@ import { MockComponents } from 'ng-mocks'
 import { CardComponent } from '../../card/card.component'
 import { CardHeaderImageComponent } from '../../card/card-header/card-header-image/card-header-image.component'
 import { CardHeaderComponent } from '../../card/card-header/card-header.component'
-import { ProjectItem, Stack } from './project-item'
+import { ProjectItem, Stack } from '../../data/project-item'
 import { CardHeaderTextsComponent } from '../../card/card-header/card-header-texts/card-header-texts.component'
 import { CardHeaderDetailComponent } from '../../card/card-header/card-header-detail/card-header-detail.component'
 import { DateRangeComponent } from '../../date-range/date-range.component'
